+++ src/PSR7Csrf/Session/ExtractUniqueKeyFromSession.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace PSR7Csrf\Session;
 
-use Lcobucci\JWT\Signer;
-use PSR7Session\Session\SessionInterface;
+use PSR7Sessions\Storageless\Session\SessionInterface;
 
 final class ExtractUniqueKeyFromSession implements ExtractUniqueKeyFromSessionInterface
 {