+++ test/PSR7CsrfTest/CSRFCheckerMiddlewareTest.php
@@ -7,22 +7,22 @@ namespace PSR7CsrfTest;
 use Lcobucci\JWT\Builder;
 use Lcobucci\JWT\Parser;
 use Lcobucci\JWT\Signer;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Message\StreamInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use PSR7Csrf\CSRFCheckerMiddleware;
 use PSR7Csrf\Exception\SessionAttributeNotFoundException;
 use PSR7Csrf\HttpMethod\IsSafeHttpRequestInterface;
 use PSR7Csrf\RequestParameter\ExtractCSRFParameterInterface;
 use PSR7Csrf\Session\ExtractUniqueKeyFromSessionInterface;
-use PSR7Session\Session\SessionInterface;
+use PSR7Sessions\Storageless\Session\SessionInterface;
 use stdClass;
 
 /**
  * @covers \PSR7Csrf\CSRFCheckerMiddleware
  */
-final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
+final class CSRFCheckerMiddlewareTest extends TestCase
 {
     /**
      * @var Signer
@@ -70,7 +70,7 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
     private $sessionAttribute;
 
     /**
-     * @var callable|\PHPUnit_Framework_MockObject_MockObject
+     * @var RequestHandlerInterface|\PHPUnit_Framework_MockObject_MockObject
      */
     private $nextMiddleware;
 
@@ -78,6 +78,11 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
      * @var CSRFCheckerMiddleware
      */
     private $middleware;
+
+    /**
+     * @var ResponseInterface
+     */
+    private $faultyResponse;
 
     /**
      * {@inheritDoc}
@@ -88,21 +93,23 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
 
         $this->signer                      = new Signer\Hmac\Sha256();
         $this->tokenParser                 = new Parser();
-        $this->isSafeHttpRequest           = $this->getMock(IsSafeHttpRequestInterface::class);
-        $this->extractUniqueKeyFromSession = $this->getMock(ExtractUniqueKeyFromSessionInterface::class);
-        $this->extractCSRFParameter        = $this->getMock(ExtractCSRFParameterInterface::class);
-        $this->request                     = $this->getMock(ServerRequestInterface::class);
-        $this->response                    = $this->getMock(ResponseInterface::class);
-        $this->session                     = $this->getMock(SessionInterface::class);
+        $this->isSafeHttpRequest           = $this->createMock(IsSafeHttpRequestInterface::class);
+        $this->extractUniqueKeyFromSession = $this->createMock(ExtractUniqueKeyFromSessionInterface::class);
+        $this->extractCSRFParameter        = $this->createMock(ExtractCSRFParameterInterface::class);
+        $this->request                     = $this->createMock(ServerRequestInterface::class);
+        $this->response                    = $this->createMock(ResponseInterface::class);
+        $this->session                     = $this->createMock(SessionInterface::class);
         $this->sessionAttribute            = uniqid('session', true);
-        $this->nextMiddleware              = $this->getMock(stdClass::class, ['__invoke']);
+        $this->nextMiddleware              = $this->createMock(RequestHandlerInterface::class);
+        $this->faultyResponse              = $this->createMock(ResponseInterface::class);
         $this->middleware                  = new CSRFCheckerMiddleware(
             $this->isSafeHttpRequest,
             $this->extractUniqueKeyFromSession,
             $this->extractCSRFParameter,
             $this->tokenParser,
             $this->signer,
-            $this->sessionAttribute
+            $this->sessionAttribute,
+            $this->faultyResponse
         );
     }
 
@@ -110,34 +117,29 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
     {
         $this->isSafeHttpRequest->expects(self::any())->method('__invoke')->with($this->request)->willReturn(true);
 
-        self::assertSame($this->response, $this->middleware->__invoke($this->request, $this->response));
+        $this
+            ->nextMiddleware
+            ->expects(self::once())
+            ->method('handle')
+            ->with($this->request)
+            ->willReturn($this->response);
+
+        self::assertSame($this->response, $this->middleware->process($this->request, $this->nextMiddleware));
     }
 
-    public function testWillIgnoreSafeRequestsWithoutNextMiddleware()
+    public function testWillSucceedIfANonSafeRequestIsProvidedWithAValidTokenWithNextMiddleware()
     {
-        $nextReturnValue = $this->getMock(ResponseInterface::class);
+        $secret          = uniqid('secret', true);
+        $validToken      = (new Builder())
+            ->sign($this->signer, $secret)
+            ->getToken();
 
         $this
             ->nextMiddleware
             ->expects(self::once())
-            ->method('__invoke')
-            ->with($this->request, $this->response)
-            ->willReturn($nextReturnValue);
-        $this->isSafeHttpRequest->expects(self::any())->method('__invoke')->with($this->request)->willReturn(true);
-
-        self::assertSame(
-            $nextReturnValue,
-            $this->middleware->__invoke($this->request, $this->response, $this->nextMiddleware)
-        );
-    }
-
-    public function testWillSucceedIfANonSafeRequestIsProvidedWithAValidTokenWithoutNextMiddleware()
-    {
-        $secret     = uniqid('secret', true);
-        $validToken = (new Builder())
-            ->sign($this->signer, $secret)
-            ->getToken();
-
+            ->method('handle')
+            ->with($this->request)
+            ->willReturn($this->response);
         $this->isSafeHttpRequest->expects(self::any())->method('__invoke')->with($this->request)->willReturn(false);
         $this
             ->extractUniqueKeyFromSession
@@ -160,47 +162,7 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
 
         self::assertSame(
             $this->response,
-            $this->middleware->__invoke($this->request, $this->response)
-        );
-    }
-
-    public function testWillSucceedIfANonSafeRequestIsProvidedWithAValidTokenWithNextMiddleware()
-    {
-        $nextReturnValue = $this->getMock(ResponseInterface::class);
-        $secret          = uniqid('secret', true);
-        $validToken      = (new Builder())
-            ->sign($this->signer, $secret)
-            ->getToken();
-
-        $this
-            ->nextMiddleware
-            ->expects(self::once())
-            ->method('__invoke')
-            ->with($this->request, $this->response)
-            ->willReturn($nextReturnValue);
-        $this->isSafeHttpRequest->expects(self::any())->method('__invoke')->with($this->request)->willReturn(false);
-        $this
-            ->extractUniqueKeyFromSession
-            ->expects(self::any())
-            ->method('__invoke')
-            ->with($this->session)
-            ->willReturn($secret);
-        $this
-            ->extractCSRFParameter
-            ->expects(self::any())
-            ->method('__invoke')
-            ->with($this->request)
-            ->willReturn((string) $validToken);
-        $this
-            ->request
-            ->expects(self::any())
-            ->method('getAttribute')
-            ->with($this->sessionAttribute)
-            ->willReturn($this->session);
-
-        self::assertSame(
-            $nextReturnValue,
-            $this->middleware->__invoke($this->request, $this->response, $this->nextMiddleware)
+            $this->middleware->process($this->request, $this->nextMiddleware)
         );
     }
 
@@ -231,7 +193,7 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
             ->with($this->sessionAttribute)
             ->willReturn($this->session);
 
-        $this->assertFaultyResponse($this->middleware, $this->request, $this->response);
+        $this->assertFaultyResponse();
     }
 
     public function testUnsignedTokensAreRejected()
@@ -259,7 +221,7 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
             ->with($this->sessionAttribute)
             ->willReturn($this->session);
 
-        $this->assertFaultyResponse($this->middleware, $this->request, $this->response);
+        $this->assertFaultyResponse();
     }
 
     public function testExpiredSignedTokensAreRejected()
@@ -290,7 +252,7 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
             ->with($this->sessionAttribute)
             ->willReturn($this->session);
 
-        $this->assertFaultyResponse($this->middleware, $this->request, $this->response);
+        $this->assertFaultyResponse();
     }
 
     public function testMalformedTokensShouldBeRejected()
@@ -315,7 +277,7 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
             ->with($this->sessionAttribute)
             ->willReturn($this->session);
 
-        $this->assertFaultyResponse($this->middleware, $this->request, $this->response);
+        $this->assertFaultyResponse();
     }
 
     public function testWillFailIfARequestDoesNotIncludeASession()
@@ -341,28 +303,13 @@ final class CSRFCheckerMiddlewareTest extends PHPUnit_Framework_TestCase
 
         $this->expectException(SessionAttributeNotFoundException::class);
 
-        $this->middleware->__invoke($this->request, $this->response, $this->nextMiddleware);
+        $this->middleware->process($this->request, $this->nextMiddleware);
     }
 
-    /**
-     * @param CSRFCheckerMiddleware                                      $middleware
-     * @param ServerRequestInterface                                     $request
-     * @param ResponseInterface|\PHPUnit_Framework_MockObject_MockObject $response
-     *
-     * @return void
-     */
-    private function assertFaultyResponse(
-        CSRFCheckerMiddleware $middleware,
-        ServerRequestInterface $request,
-        ResponseInterface $response
-    ) {
-        $faultyResponse = $this->getMock(ResponseInterface::class);
-        $responseBody   = $this->getMock(StreamInterface::class);
+    private function assertFaultyResponse() : void
+    {
+        $this->nextMiddleware->expects(self::never())->method('handle');
 
-        $response->expects(self::any())->method('withStatus')->with(400)->willReturn($faultyResponse);
-        $faultyResponse->expects(self::any())->method('getBody')->willReturn($responseBody);
-        $responseBody->expects(self::once())->method('write')->with('{"error":"missing or invalid CSRF token"}');
-
-        self::assertSame($faultyResponse, $middleware->__invoke($request, $response));
+        self::assertSame($this->faultyResponse, $this->middleware->process($this->request, $this->nextMiddleware));
     }
 }