+++ test/PSR7CsrfTest/Exception/SessionAttributeNotFoundExceptionTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PSR7CsrfTest\Exception;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ServerRequestInterface;
 use PSR7Csrf\Exception\ExceptionInterface;
 use PSR7Csrf\Exception\SessionAttributeNotFoundException;
@@ -13,12 +13,12 @@ use UnexpectedValueException;
 /**
  * @covers \PSR7Csrf\Exception\SessionAttributeNotFoundException
  */
-final class SessionAttributeNotFoundExceptionTest extends PHPUnit_Framework_TestCase
+final class SessionAttributeNotFoundExceptionTest extends TestCase
 {
     public function testFromInvalidExpirationTime()
     {
         /* @var $request ServerRequestInterface|\PHPUnit_Framework_MockObject_MockObject */
-        $request = $this->getMock(ServerRequestInterface::class);
+        $request = $this->createMock(ServerRequestInterface::class);
 
         $request->expects(self::any())->method('getAttributes')->willReturn(['foo' => 'bar', 'baz' => 'tab']);