+++ test/PSR7CsrfTest/Session/ExtractUniqueKeyFromSessionTest.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace PSR7CsrfTest\HttpMethod;
 
-use Lcobucci\JWT\Signer;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use PSR7Csrf\Session\ExtractUniqueKeyFromSession;
-use PSR7Session\Session\SessionInterface;
+use PSR7Sessions\Storageless\Session\SessionInterface;
 
 /**
  * @covers \PSR7Csrf\Session\ExtractUniqueKeyFromSession
  */
-final class ExtractUniqueKeyFromSessionTest extends PHPUnit_Framework_TestCase
+final class ExtractUniqueKeyFromSessionTest extends TestCase
 {
     /**
      * @dataProvider keysProvider
@@ -22,7 +21,7 @@ final class ExtractUniqueKeyFromSessionTest extends PHPUnit_Framework_TestCase
     public function testExtractionWithExistingKey(string $key)
     {
         /* @var $session SessionInterface|\PHPUnit_Framework_MockObject_MockObject */
-        $session     = $this->getMock(SessionInterface::class);
+        $session     = $this->createMock(SessionInterface::class);
         $superSecret = uniqid('', true);
 
         $session->expects(self::any())->method('get')->with($key, '')->willReturn($superSecret);
@@ -41,7 +40,7 @@ final class ExtractUniqueKeyFromSessionTest extends PHPUnit_Framework_TestCase
         $extractKey = new ExtractUniqueKeyFromSession($key);
 
         /* @var $session SessionInterface|\PHPUnit_Framework_MockObject_MockObject */
-        $session = $this->getMock(SessionInterface::class);
+        $session = $this->createMock(SessionInterface::class);
 
         $session->expects(self::any())->method('get')->with($key, '')->willReturn('');
         $session->expects(self::exactly(2))->method('set')->with(
@@ -75,7 +74,7 @@ final class ExtractUniqueKeyFromSessionTest extends PHPUnit_Framework_TestCase
         $extractKey = new ExtractUniqueKeyFromSession($key);
 
         /* @var $session SessionInterface|\PHPUnit_Framework_MockObject_MockObject */
-        $session = $this->getMock(SessionInterface::class);
+        $session = $this->createMock(SessionInterface::class);
 
         $session->expects(self::any())->method('get')->with($key, '')->willReturn(123);
         $session->expects(self::exactly(2))->method('set')->with(